remove cordova util file
